requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==6.6
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.6.1
requests==2.23.0
soupsieve==2.0
urllib3==1.25.8
uWSGI==2.0.14
Werkzeug==1.0.0
PyJWT==1.7.1